+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 New functionality that are backwards-compatible:
 * [OLMIS-7604](https://openlmis.atlassian.net/browse/OLMIS-7604): Added valid destination page to the administration panel
-
+* [OLMIS-7603](https://openlmis.atlassian.net/browse/OLMIS-7603): Added valid sources in administration panel
 
 5.6.7 / 2022-04-21
 ==================
@@ -14,6 +14,7 @@ New functionality that are backwards-compatible:
 * [OLMIS-7444](https://openlmis.atlassian.net/browse/OLMIS-7444): Add edit modal to lot management page
 * [OLMIS-7452](https://openlmis.atlassian.net/browse/OLMIS-7452): Add ORDER_CREATE right
 * [OLMIS-7501](https://openlmis.atlassian.net/browse/OLMIS-7501): Add CCE_INVENTORY_TRANSFER right
+
 Improvements:
 * [OLMIS-6545](https://openlmis.atlassian.net/browse/OLMIS-6545): Add missing translation adminOrderableEdit.failedToUpdateProduct